Chapter 10 Lab
This lab entails creating neural networks from scratch with class functions and using prebuilt functions. It goes over single and multilayer neural networks and then builds to convolutional and recurrent neural networks. It compares there results with logistic and linear regression and lasso. Finally, it uses data visualization techniques to show the effects of different parameters on the neural network.