Chapter 13 Notes
Notes on different resampling methods, False Discovery Rate equations , Family wise error rate, Tukey, scheffe, Holms, and Bonferroni equations. 

The goal of this chapter was to test multiple hypothesis tests on huge data sets. Huge data plays a risk in analyzing the p value. So we went over ways to analyze this in a safe way. Starting with Family Wise Error Rate, then moving to algorithms to better manage that like Bonferroni, Holms, Scheffesand Tukey. Then realizing that these equations are too strict with huge values of M. So we learned about False Discovery Rate and how to calculate that using Benjamin- Hochberg. Lastly, we went over the issues with lack of sampling and distribution, and how to fix that using resmapling.