Chapter 13 Applied
Finished working on question 8. We created a fake data set and then performed several hypotheses testing metrics on it. We performed the normal p_test, Family Wise Error Rate using Bonferroni, False Discovery Rate, and the concerns of cherry-picking data. 

This is the last assignment of the book. It's been great fun, onto elements of statistical learning and maybe a job XD!